Add files via upload
This portfolio website serves as a showcase for my skills and projects in the field of AI and machine learning. It features a clean layout with dedicated sections for my resume, various projects, and contact information. Each project is presented with images, detailed descriptions, and links to live demos or GitHub repositories. The site is fully responsive, ensuring a seamless experience across devices, and highlights my commitment to both design and functionality.

Technologies Used:

HTML
CSS
Flexbox for layout
Responsive design principles